feat: 토큰 기반 SecurityContextInterceptor 추상화 - TokenSecurityContextHolderStrategy 추가
