+++ src/main/java/nextstep/auth/authorization/TokenSecurityContextHolderStrategy.java
@@ -1,0 +1,40 @@
+package nextstep.auth.authorization;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.auth.context.Authentication;
+import nextstep.auth.context.SecurityContext;
+import nextstep.auth.token.JwtTokenProvider;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.Map;
+
+public class TokenSecurityContextHolderStrategy implements SecurityContextHolderStrategy {
+    private final JwtTokenProvider jwtTokenProvider;
+
+    public TokenSecurityContextHolderStrategy(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    @Override
+    public SecurityContext extract(HttpServletRequest request) {
+        String credentials = AuthorizationExtractor.extract(request, AuthorizationType.BEARER);
+        if (!jwtTokenProvider.validateToken(credentials)) {
+            return null;
+        }
+        return extractSecurityContext(credentials);
+    }
+
+    private SecurityContext extractSecurityContext(String credentials) {
+        try {
+            String payload = jwtTokenProvider.getPayload(credentials);
+            TypeReference<Map<String, String>> typeRef = new TypeReference<Map<String, String>>() {
+            };
+
+            Map<String, String> principal = new ObjectMapper().readValue(payload, typeRef);
+            return new SecurityContext(new Authentication(principal));
+        } catch (Exception e) {
+            return null;
+        }
+    }
+}
